[Fix]Diet 데이터 캐싱 키값 수정
